# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PubKeyMail API
  version: 1.0.0
  description: >-
    Blockchain-based email service with wallet authentication.


    ## Authentication

    PubKeyMail uses wallet-based authentication. Users sign a challenge message 

    with their blockchain wallet and receive a JWT token for subsequent
    requests.


    ## Case Sensitivity

    **CRITICAL**: All blockchain addresses are case-sensitive. Different casing 

    represents different addresses.
  contact:
    name: PubKeyMail Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pubkeymail.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/verify endpoint
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: array
              items: {}
          required:
            - code
            - message
      required:
        - error
...
    StripeCheckoutRequest:
      type: object
      properties:
        userId:
          type: number
          exclusiveMinimum: 0
        plan:
          type: string
          enum:
            - monthly
            - yearly
        successUrl:
          type: string
          format: uri
        cancelUrl:
          type: string
          format: uri
      required:
        - userId
        - plan
        - successUrl
        - cancelUrl
    StripeCheckoutResponse:
      type: object
      properties:
        sessionId:
          type: string
          example: cs_test_...
        checkoutUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
      required:
        - sessionId
        - checkoutUrl
        - expiresAt
    SolanaPayRequest:
      type: object
      properties:
        userId:
          type: number
          exclusiveMinimum: 0
        plan:
          type: string
          enum:
            - monthly
            - yearly
      required:
        - userId
        - plan
    SolanaPayResponse:
      type: object
      properties:
        reference:
          type: string
        recipient:
          type: string
        amount:
          type: number
        splToken:
          type: string
        label:
          type: string
        message:
          type: string
        memo:
          type: string
      required:
        - reference
        - recipient
        - amount
        - splToken
        - label
        - message
        - memo
    SolanaPayVerifyRequest:
      type: object
      properties:
        reference:
          type: string
        signature:
          type: string
      required:
        - reference
        - signature
    CancelSubscriptionRequest:
      type: object
      properties:
        userId:
          type: number
          exclusiveMinimum: 0
      required:
        - userId
...
paths:
...
  /api/v1/payments/stripe/checkout:
    post:
      tags:
        - Payments
      summary: Create Stripe checkout session
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StripeCheckoutRequest"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeCheckoutResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Stripe not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/payments/solana-pay/request:
    post:
      tags:
        - Payments
      summary: Create Solana Pay request
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolanaPayRequest"
      responses:
        "200":
          description: Payment request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolanaPayResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Solana Pay not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/payments/solana-pay/verify:
    post:
      tags:
        - Payments
      summary: Verify Solana Pay transaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolanaPayVerifyRequest"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  paymentId:
                    type: string
                required:
                  - verified
        "400":
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/payments/webhooks/stripe:
    post:
      tags:
        - Payments
      summary: Stripe webhook handler
      parameters:
        - schema:
            type: string
          required: true
          name: stripe-signature
          in: header
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
              additionalProperties: {}
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                required:
                  - received
        "400":
          description: Invalid signature or webhook error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/payments/cancel:
    post:
      tags:
        - Payments
      summary: Cancel subscription
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelSubscriptionRequest"
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Subscription cancelled successfully
                required:
                  - message
        "400":
          description: Cancellation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...